backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.36
alembic==1.13.3
asyncpg==0.30.0
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
httpx==0.27.2
groq==0.11.0
anthropic==0.34.2
boto3==1.35.32
resend==2.3.0
apscheduler==3.10.4
python-dotenv==1.0.1
email-validator==2.2.0
